Add AlchemyCMS content provider
This allows to load content from an Alchemy page. It also allows to filter the elements to include with the page, making it possible to only load a specific element from a page.

Make sure to have `gem 'alchemy_cms'` added to your `Gemfile`.

It was necessary to include an inititalizer and test setup code. I tried my best to make it as unobstrusive as possible.